[BUGFIX] IndexManager: Check if index exists before deleting
Leads to a ES error otherwise

+++ Classes/IndexManager.php
@@ -50,7 +50,10 @@ class IndexManager implements SingletonInterface {
             'index' => $index
         ];
 
-        $response = $this->client->indices()->delete($deleteParams);
+        if ($this->client->indices()->exists($deleteParams)) {
+
+            $response = $this->client->indices()->delete($deleteParams);
+        }
 
         $params = [
             'index' => $index,
